requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
aiosignal==1.2.0
APScheduler==3.8.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
charset-normalizer==2.0.8
colorlog==6.4.1
frozenlist==1.2.0
hikari==2.0.0.dev103
hikari-lightbulb==2.0.1
hikari-sake @ git+https://github.com/FasterSpeeding/Sake@630a542c695eb15619f3cb20d8baea8e070b1cd7
hikari-yuyo==1.0.4a1
hiredis==2.0.0
idna==3.3
multidict==5.2.0
pure25519==0.0.1
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
uvloop==0.16.0
yarl==1.7.2
